Implements control signal handling and Flash storage
Introduces a ControlSignalHandler class to manage limit switches and signal handling.
This allows registering a callback function that is executed when a limit switch is triggered.

Implements FlashController to handle persistent storage of configuration data using EEPROM.
Adds functionality for reading and writing pin configurations and version control to manage data migrations.

Refactors LedControl to use FlashController for storing the LED pin number.